Clean up code and mount dashboard on /dashboard to prevent middleware overspill
