+++ app.js
@@ -1,9 +1,6 @@
 'use strict';
 const express = require('express');
-const router = express.Router();
 const logger = require('morgan');
-const nunjucks = require('nunjucks');
-const moment = require('moment');
 
 const env = process.env.NODE_ENV || 'development';
 if (env === 'development') {
@@ -12,47 +9,29 @@ if (env === 'development') {
 const passportMiddleware = require('./services/passport');
 const sessionMiddleware = require('./services/session');
 const bodyParserMiddleware = require('./services/bodyParser');
+const viewEngine = require('./services/viewEngine');
 
 const app = express();
+
+// Set up templating
+app.set('view engine', 'njk').set('engineEnv', viewEngine(app));
+
+// Configure middleware
+app.use(logger('dev'));
 app.use(bodyParserMiddleware);
 app.use(sessionMiddleware(app));
 app.use(passportMiddleware());
-app.use(logger('dev'));
 
-const templateEnv = nunjucks.configure('.', {
-    autoescape: true,
-    express: app
-});
-
-/**
- * View helper for formatting a date
- * @param {String} dateString
- * @param {String} format
- * @see https://momentjs.com/docs/#/displaying/format/
- */
-templateEnv.addFilter('formatDate', function(dateString, format) {
-    return moment(dateString).format(format);
-});
-
-/**
- * View helper to represent date as relative time
- * @param {String} dateString
- */
-templateEnv.addFilter('timeFromNow', function(dateString) {
-    return moment(dateString).fromNow();
-});
-
-app.set('view engine', 'njk').set('engineEnv', templateEnv);
-
-app.use('/', require('./apps/dashboard'));
-// router.use('/api', require('./apps/api'));
+// Mount routes
+app.use('/dashboard', require('./apps/dashboard'));
+app.use('/api', require('./apps/api'));
 app.use('/user', require('./apps/user'));
 
+// Handle errors
 app.use((err, req, res, next) => {
     if (res.headersSent) {
         return next(err);
     }
-
     res.send({err, req, res});
 });
